fix dlsym
