+++ internal/ossl/ossl_nocgo.go
@@ -6,17 +6,6 @@ import (
 	"unsafe"
 )
 
-func dlsym(handle unsafe.Pointer, symbol string, optional bool) unsafe.Pointer {
-	r0 := Dlsym(handle, unsafe.StringData(symbol))
-	if r0 == nil {
-		if !optional {
-			panic("cannot get required symbol " + symbol)
-		}
-		return nil
-	}
-	return unsafe.Pointer(r0)
-}
-
 // HashSum copies ctx1 into ctx2 and calls EVP_DigestFinal_ex using ctx2.
 // This is necessary because Go hash.Hash mandates that Sum has no effect
 // on the underlying stream. In particular it is OK to Sum, then Write more,
